requirements.txt

# Django
django~=3.0.5
elasticsearch-dsl==7.3.0
djangorestframework~=3.11.0

# Django Configuration
django-environ~=0.4.5